VehicleOOP.py
"""Create a vehicle that comprises different parts that can be selected to be added to the vehicle. 
Each vehicle has its own serial number and name. Each vehicle can only have up to three wheels. 
For all parts, keep a note of the type of part, its serial number and an age of the part. For wheels, store their diameter and material. 
For normal parts, they can be determined to be old if they are over two years old. For wheels, if the wheel is above 1m in diameter, it is old after a year, otherwise it is old after three years."""